+++ library/src/iusSourceDict.c
@@ -215,7 +215,7 @@ iusd_t iusSourceDictLoad
     status = H5Gget_num_objs(grpid, &nobj);
 
     iusd_t dict = iusSourceDictCreate();
-    for (i = 0; i < nobj; i++)
+    for (i = 0; i < (int) nobj; i++)
     {
         H5Gget_objname_by_idx(grpid, (hsize_t) i,
                                     memb_name, (size_t) MAX_NAME);
